use floats everywhere for physics

+++ cotta-core/src/main/kotlin/com/mgtriffid/games/cotta/core/clock/CottaClock.kt
@@ -2,5 +2,5 @@ package com.mgtriffid.games.cotta.core.clock
 
 interface CottaClock {
     fun time(): Long
-    fun delta(): Double
+    fun delta(): Float
 }
