+++ panna-shared/src/main/kotlin/com/mgtriffid/games/panna/shared/game/components/physics/VelocityComponent.kt
@@ -5,20 +5,20 @@ import com.mgtriffid.games.cotta.core.entities.MutableComponent
 
 interface VelocityComponent : MutableComponent<VelocityComponent> {
     @ComponentData
-    var velX: Int
+    var velX: Float
     @ComponentData
-    var velY: Int
+    var velY: Float
 
     companion object {
-        fun create(velX: Int, velY: Int): VelocityComponent {
+        fun create(velX: Float, velY: Float): VelocityComponent {
             return VelocityComponentImpl(velX, velY)
         }
     }
 }
 
 private data class VelocityComponentImpl(
-    override var velX: Int,
-    override var velY: Int
+    override var velX: Float,
+    override var velY: Float
 ) : VelocityComponent {
     override fun copy(): VelocityComponent {
         return this.copy(velX = velX, velY = velY)
